requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.2
idna==2.9
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
requests==2.23.0
selenium==3.141.0
six==1.14.0
traitlets==4.3.3
urllib3==1.25.8
validators==0.14.3
wcwidth==0.1.9
